home v8

+++ Crud/follow_crud.py
@@ -1,4 +1,6 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException
+from . import user_crud
 from .. import models
 
 def check_following(db:Session,user_id: int, photographer_id: int):
@@ -6,15 +8,29 @@ def check_following(db:Session,user_id: int, photographer_id: int):
     user_exists = db.query(exists_query).scalar()
     return user_exists
 
-# def is_photographer(db: Session,photographer_id: int):
+def is_photographer(db: Session,photographer_id: int):
+    try:
+        photographer=user_crud.get_user_by_userid(db,photographer_id)
+        return photographer.user_type == 'photographer'
+    except :
+        raise HTTPException(status_code=401, detail="No such photographer ID exists")
 
-# def is_simple_user(db: Session,photographer_id: int):
+
+def is_simple_user(db: Session,user_id: int):
+    simple=user_crud.get_user_by_userid(db,user_id)
+    return simple.user_type == 'simple'
 
 def follow_photographer(db: Session, user_id: int, photographer_id: int):
+    if not is_photographer(db,photographer_id):
+        raise HTTPException(status_code=401, detail="Please enter only photographer ID")
+    
+    if not is_simple_user(db,user_id):
+        raise HTTPException(status_code=401, detail="Only simple user can follow/unfollow a photographer")
+        
     exist=check_following(db,user_id,photographer_id)
     if exist:
         return None
-    # check whether the photographer_id belongs to photographer and same for user
+    
     follow = models.Follow(user_id=user_id, photographer_id=photographer_id)
     db.add(follow)
     db.commit()
@@ -22,10 +38,17 @@ def follow_photographer(db: Session, user_id: int, photographer_id: int):
     return follow
 
 def unfollow_photographer(db: Session, user_id: int, photographer_id: int):
+
+    if not is_photographer(db,photographer_id):
+        raise HTTPException(status_code=401, detail="Please enter correct/existing photograher ID")
+    
+    if not is_simple_user(db,user_id):
+        raise HTTPException(status_code=401, detail="Only Simple user can follow/unfollow a photographer")
+
     exist=check_following(db,user_id,photographer_id)
     if not exist:
         return None
-     # check weather the photographer_id belongs to photographer and same for user
+
     follow = db.query(models.Follow).filter_by(user_id=user_id, photographer_id=photographer_id).first()
     if follow:
         db.delete(follow)
